+++ Source/Program.cs
@@ -8,6 +8,14 @@ bld.Services
            c.DefaultRequestHeaders.AcceptLanguage.ParseAdd("en-US,en;q=0.9");
            c.DefaultRequestHeaders.UserAgent.ParseAdd(
                "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36 Edg/130.0.0.0");
+           c.Timeout = TimeSpan.FromSeconds(10);
+       }).Services
+   .AddHttpClient(
+       "pushbullet",
+       c =>
+       {
+           c.DefaultRequestHeaders.Add("Access-Token", bld.Configuration["Pushbullet:ApiKey"]);
+           c.Timeout = TimeSpan.FromSeconds(10);
        }).Services
    .AddOutputCache()
    .AddFastEndpoints();